Consistencia y warnings

+++ Userland/syscalls/syscalls.c
@@ -65,15 +65,15 @@ void sys_set_screensaver_timer(uint64_t t) {
 	syscall((uint64_t)SYSCALL_SCREENSAVER_TIMER, t);
 }
 
-void sys_clear_screen(){
-	syscall(SYSCALL_CLEAR_SCREEN);
+void sys_clear_screen() {
+	syscall((uint64_t)SYSCALL_CLEAR_SCREEN);
 }
 
 void sys_screensaver_trigger() {
 	syscall((uint64_t) SYSCALL_SCREENSAVER_TRIGGER);
 }
 
-void sys_exit(){
-	syscall(SYSCALL_EXIT);
+void sys_exit() {
+	syscall((uint64_t)SYSCALL_EXIT);
 }
